DTO/Service/Repository/Controller Long→Integer 추가 수정 (3차)
- AccommodationCreateRequestDTO/UpdateRequestDTO: mainRegionId, subRegionId Long→Integer
- dtoOwner/AccommodationDTO: mainRegionId, subRegionId Long→Integer
- OCompanyService: removeCompany(Long)→Integer
- ORegionService: getRegion(Long)→Integer
- OReservationRepository: company 관련 쿼리 파라미터 Long→Integer
- AccommodationServiceImpl: 로컬변수 Long mainRegionId/subRegionId→Integer
- CompanyAPIController: @PathVariable Long id→Integer
- ReviewController: companyId.longValue() 캐스트로 컴파일 오류 해결

+++ src/main/java/com/honeyrest/honeyrest_host/controllerAdmin/CompanyAPIController.java
@@ -35,19 +35,19 @@ public class CompanyAPIController {
 
     @Operation(summary = "업체 단건 조회")
     @GetMapping("/{id}")
-    public CompanyDTO get(@PathVariable Long id) {
+    public CompanyDTO get(@PathVariable Integer id) {
         return companyService.getById(id);
     }
 
     @Operation(summary = "업체 수정(전체/부분)")
     @PutMapping("/{id}")
-    public CompanyDTO update(@PathVariable Long id, @RequestBody CompanyDTO req) {
+    public CompanyDTO update(@PathVariable Integer id, @RequestBody CompanyDTO req) {
         return companyService.update(id, req);
     }
 
     @Operation(summary = "업체 삭제")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         companyService.delete(id);
         return ResponseEntity.noContent().build();
     }
